requirements.txt
torch==1.9.1
torchvision==0.10.1
matplotlib==3.4.3
Pillow==8.4.0
